+++ TrelloDotNet/TrelloDotNet/Model/Webhook/WebhookActionData.cs
@@ -77,5 +77,7 @@ namespace TrelloDotNet.Model.Webhook
         [JsonInclude]
         public WebhookActionDataMember Member { get; set; }
 
+        public WebhookAction Parent { get; internal set; }
+
     }
 }