+++ TrelloDotNet/TrelloDotNet/Model/Webhook/WebhookActionDataCard.cs
@@ -1,4 +1,5 @@
 ﻿using System.Text.Json.Serialization;
+using System.Threading.Tasks;
 
 namespace TrelloDotNet.Model.Webhook
 {
@@ -34,5 +35,19 @@ namespace TrelloDotNet.Model.Webhook
         [JsonPropertyName("dueComplete")]
         [JsonInclude]
         public bool? DueComplete { get; private set; }
+
+        /// <summary>
+        /// Get the Full Card Object
+        /// </summary>
+        /// <returns>The Card</returns>
+        public async Task<Card> GetAsync()
+        {
+            return await Parent.Parent.TrelloClient.GetCardAsync(Id);
+        }
+
+        /// <summary>
+        /// Parent
+        /// </summary>
+        public WebhookActionData Parent { get; internal set; }
     }
 }