+++ Changelog.md
@@ -1,6 +1,9 @@
 # Changelog: 
 *Below is the version history of [TrelloDotNet](https://github.com/rwjdk/TrelloDotNet) (An wrapper of the Trello API)*
 
+## Unreleased
+- WebhookAction now have reference to the TrelloClient and the sub-objects can get their Full Objects
+
 ## 1.2.1 (25rd of Feb. 2023)
 - Added: Generic WebHookNotification that does not care if Webhook returned from a Board, List, Card, etc
 - Added: Added Method 'UpdateCommentActionAsync' for the ability to update Comments